+++ com.unity.render-pipelines.universal/Runtime/Passes/CapturePass.cs
@@ -29,17 +29,15 @@ namespace UnityEngine.Rendering.Universal
         /// <inheritdoc/>
         public override void Execute(ScriptableRenderContext context, ref RenderingData renderingData)
         {
-            CommandBuffer cmdBuf = CommandBufferPool.Get();
+            CommandBuffer cmdBuf = renderingData.commandBuffer;
+
             using (new ProfilingScope(cmdBuf, m_ProfilingSampler))
             {
                 var colorAttachmentIdentifier = m_CameraColorHandle.nameID;
                 var captureActions = renderingData.cameraData.captureActions;
                 for (captureActions.Reset(); captureActions.MoveNext();)
-                    captureActions.Current(colorAttachmentIdentifier, cmdBuf);
+                    captureActions.Current(colorAttachmentIdentifier, renderingData.commandBuffer);
             }
-
-            context.ExecuteCommandBuffer(cmdBuf);
-            CommandBufferPool.Release(cmdBuf);
         }
     }
 }